Cluster: I have add a method for adding to its set of precincts and also to increase the population
Demographics:
Created a demographics class to hold all the percentages

State:
Added a few methods to help with breakCluster and use case 29.
toDistrict which ensures that there is the correct number of districts and if not to break apart the smallest cluster.
Also added eligibleCluster to determine where a given precinct should be sent based on population.
Minpopulation is helpful because the clusters are stored in a set, but i need to constantly find the the cluster with the min population to receive the new precinct.